fix: add intermittent cleanup in bulk tests
As it turns out, GKE has a default limit of 100
Ingress resources, which we hit with our bulk test
immediately. This patch causes each test step that
deploys bulk resources to clean up after itself,
and reduces the overally number of Ingress resources
that will ever be deployed during the test.